Delete all pseudocode to start again
The pseudocode was getting too detailed which I think went against
the original idea of pseudocode to help with structural planning,
not to rewrite the entire logic of the program in English.

I will write the pseudocode again, this time I will do it faster and be
less detailed. I will save the details for the actual coding.